refactor: extract getEditModerationController into its own file
